+++ src/turboquant_cpp/polar_quant.h
@@ -1,8 +1,6 @@
 #pragma once
-// PolarQuant block-level and tensor-level compress/decompress pipeline.
-//
-// Pipeline: norm extraction → WHT rotation → scalar quantization → packing
-// Polar decomposition separates magnitude (norm) from direction (unit vector).
+// PolarQuant compression pipeline (float precision).
+// Uses thread-local scratch buffers to eliminate per-block allocations.
 
 #include <cstdint>
 #include <vector>
@@ -21,30 +19,27 @@ struct CompressedBlock {
 /// Compressed tensor: collection of blocks.
 struct CompressedTensor {
     std::vector<CompressedBlock> blocks;
-    std::vector<int> original_shape;  // Shape before compression
+    std::vector<int> original_shape;
     int n_bits;
     int block_size;
 };
 
-/// Compress a single block using PolarQuant pipeline.
-/// Pipeline: norm → normalize → pad → rotate → scale → quantize
-CompressedBlock polar_quantize_block(const double* data, int len,
+/// Compress a single block.  Uses thread-local scratch buffers.
+CompressedBlock polar_quantize_block(const float* data, int len,
                                      int n_bits, int seed);
 
-/// Decompress a single CompressedBlock.
-/// Pipeline: dequantize → unscale → inverse_rotate → trim → rescale
-std::vector<double> polar_dequantize_block(const CompressedBlock& block);
+/// Decompress a single block.  Uses thread-local scratch buffers.
+std::vector<float> polar_dequantize_block(const CompressedBlock& block);
 
-/// Compress a full tensor (flat array) with blocking.
-/// Block i uses seed = base_seed + i.
-CompressedTensor polar_quantize(const double* data, int total_elements,
+/// Compress full tensor with blocking + OpenMP.
+CompressedTensor polar_quantize(const float* data, int total_elements,
                                 const int* shape, int ndim,
                                 int n_bits, int seed, int block_size);
 
-/// Decompress a CompressedTensor back to flat array.
-std::vector<double> polar_dequantize(const CompressedTensor& ct);
+/// Decompress full tensor.
+std::vector<float> polar_dequantize(const CompressedTensor& ct);
 
-/// Compression ratio vs float16 storage.
+/// Compression ratio vs float16.
 double compression_ratio(int n_bits, int block_size);
 
 }  // namespace turboquant
